Added scrolling to the canvas, intervals, fixed domains
Added to argparser to kinda cheat/fix x**x and x**(1/2) x**0.5. This obviously does not really fix the issue we need to fix it by making the parser able to figure out the domain. I added scrolling to canvas which does not exactly work great but alongside the ability to choose interval adds some more functionality.